Grpc/requirements.txt
Flask
requests
grpcio
grpcio-tools
protobuf
pika
pymssql
features